+++ src/layout/index.js
@@ -1,12 +1,13 @@
 export { default as dimensionsHeatTable } from "./heatTableDimensions.js";
 export { default as dimensionsHorizontalBarChart } from "./horizontalBarChartDimensions.js";
-export { default as dimensionsVerticalBarChart } from "./verticalBarChartDimensions.js";
+export * from "./legendColorOrdinalLayout.js";
 export { default as layoutPopulationPyramid } from "./populationPyramidLayout.js";
+export { computeLayout as sankeyLayout, prepareData as sankeyPrepareData } from "./sankey.js";
 export { default as layoutSmallMultiples } from "./smallMultiples.js";
 export { default as layoutStackedAreaMultiples } from "./stackedAreaMultiplesLayout.js";
-export { prepareData as sankeyPrepareData, computeLayout as sankeyLayout } from "./sankey.js";
 export {
-  prepareData as sunburstPrepareData,
   computeLayout as sunburstLayout,
-  getRadiusExtent as sunburstGetRadiusExtent
+  getRadiusExtent as sunburstGetRadiusExtent,
+  prepareData as sunburstPrepareData
 } from "./sunburst.js";
+export { default as dimensionsVerticalBarChart } from "./verticalBarChartDimensions.js";